requirements.txt
aiosqlite==0.20.0
anyio==4.3.0
certifi==2024.2.2
exceptiongroup==1.2.1
Faker==25.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
six==1.16.0
sniffio==1.3.1
typing_extensions==4.11.0
